fix: Use temporary file correctly
By utilizing a NamedTemporaryFile, we can temporarily save the image to 
disk so it can be used in the deepfry process and Discord's File class 
will still read it as if it was a normal file.

The `with` context manager will ensure it the file gets cleaned off disk 
once it exits the context.

Also add description to url param for fry_to_shits while adding docs for 
new temp_file param.